some indent fixes from Martin Paljak plus some additional changes from me
git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@2083 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/pkcs11/misc.c
@@ -24,9 +24,9 @@
 
 #define DUMP_TEMPLATE_MAX	32
 
-void strcpy_bp(u8 *dst, const char *src, int dstsize)
+void strcpy_bp(u8 *dst, const char *src, size_t dstsize)
 {
-	int c;
+	size_t c;
 
 	if (!dst || !src || !dstsize)
 		return;
@@ -88,31 +88,31 @@ CK_RV pool_initialize(struct sc_pkcs11_pool *pool, int type)
 	pool->num_items = 0;
 	pool->head = pool->tail = NULL;
 
-        return CKR_OK;
+	return CKR_OK;
 }
 
 CK_RV pool_insert(struct sc_pkcs11_pool *pool, void *item_ptr, CK_ULONG_PTR pHandle)
 {
 	struct sc_pkcs11_pool_item *item;
-        int handle = pool->next_free_handle++;
+	int handle = pool->next_free_handle++;
 
 	item = (struct sc_pkcs11_pool_item*) malloc(sizeof(struct sc_pkcs11_pool_item));
 
 	if (pHandle != NULL)
-                *pHandle = handle;
+		*pHandle = handle;
 
-        item->handle = handle;
+	item->handle = handle;
 	item->item = item_ptr;
 	item->next = NULL;
-        item->prev = pool->tail;
+	item->prev = pool->tail;
 
 	if (pool->head != NULL && pool->tail != NULL) {
 		pool->tail->next = item;
-                pool->tail = item;
+		pool->tail = item;
 	} else
-                pool->head = pool->tail = item;
-
-        return CKR_OK;
+		pool->head = pool->tail = item;
+		
+	return CKR_OK;
 }
 
 CK_RV pool_find(struct sc_pkcs11_pool *pool, CK_ULONG handle, void **item_ptr)
@@ -120,12 +120,12 @@ CK_RV pool_find(struct sc_pkcs11_pool *pool, CK_ULONG handle, void **item_ptr)
 	struct sc_pkcs11_pool_item *item;
 
 	if (context == NULL)
-                return CKR_CRYPTOKI_NOT_INITIALIZED;
+		return CKR_CRYPTOKI_NOT_INITIALIZED;
 
 	for (item = pool->head; item != NULL; item = item->next) {
 		if (item->handle == handle) {
 			*item_ptr = item->item;
-                        return CKR_OK;
+				return CKR_OK;
 		}
 	}
 
@@ -148,7 +148,7 @@ CK_RV pool_find_and_delete(struct sc_pkcs11_pool *pool, CK_ULONG handle, void **
                         if (pool->tail == item) pool->tail = item->prev;
 
 			*item_ptr = item->item;
-                        free(item);
+			free(item);
 
 			return CKR_OK;
 		}
@@ -167,7 +167,7 @@ CK_RV session_start_operation(struct sc_pkcs11_session *session,
 	sc_pkcs11_operation_t *op;
 
 	if (context == NULL)
-                return CKR_CRYPTOKI_NOT_INITIALIZED;
+		return CKR_CRYPTOKI_NOT_INITIALIZED;
 
 	if (type < 0 || type >= SC_PKCS11_OPERATION_MAX)
 		return CKR_ARGUMENTS_BAD;
@@ -182,7 +182,7 @@ CK_RV session_start_operation(struct sc_pkcs11_session *session,
 	if (operation)
 		*operation = op;
 
-        return CKR_OK;
+	return CKR_OK;
 }
 
 CK_RV session_get_operation(struct sc_pkcs11_session *session, int type,
@@ -199,7 +199,7 @@ CK_RV session_get_operation(struct sc_pkcs11_session *session, int type,
 	if (operation)
 		*operation = op;
 
-        return CKR_OK;
+	return CKR_OK;
 }
 
 CK_RV session_stop_operation(struct sc_pkcs11_session *session, int type)
@@ -211,7 +211,7 @@ CK_RV session_stop_operation(struct sc_pkcs11_session *session, int type)
 		return CKR_OPERATION_NOT_INITIALIZED;
 
 	sc_pkcs11_release_operation(&session->operation[type]);
-        return CKR_OK;
+	return CKR_OK;
 }
 
 CK_RV attr_extract(CK_ATTRIBUTE_PTR pAttr, void *ptr, size_t *sizep)
